+++ tesseract/tesseract_collision/include/tesseract_collision/test_suite/collision_sphere_sphere_cast_unit.hpp
@@ -178,31 +178,34 @@ inline void runTestPrimitive(ContinuousContactManager& checker)
   if (result_vector[0].link_names[0] != "sphere_link")
     idx = { 1, 0, -1 };
 
-  EXPECT_NEAR(result_vector[0].cc_time[idx[0]], 0.5, 0.001);
-  EXPECT_NEAR(result_vector[0].cc_time[idx[1]], 0.5, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[0])], 0.5, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[1])], 0.5, 0.001);
 
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[0])] == ContinuousCollisionType::CCType_Between);
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[0])] == ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[0]))] ==
+              ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[0]))] ==
+              ContinuousCollisionType::CCType_Between);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.05, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.05, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.05, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], -0.05, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][0], 0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][2], 0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][0], -0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][0], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][0], -0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_TRUE(result_vector[0].transform[idx[0]].isApprox(location_start["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].transform[idx[1]].isApprox(location_start["sphere1_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[0]].isApprox(location_end["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[1]].isApprox(location_end["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[0])].isApprox(location_start["sphere_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[1])].isApprox(location_start["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].cc_transform[static_cast<size_t>(idx[0])].isApprox(location_end["sphere_link"], 0.0001));
+  EXPECT_TRUE(
+      result_vector[0].cc_transform[static_cast<size_t>(idx[1])].isApprox(location_end["sphere1_link"], 0.0001));
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
   EXPECT_NEAR(result_vector[0].normal[1], idx[2] * 0.0, 0.001);
@@ -246,31 +249,34 @@ inline void runTestPrimitive(ContinuousContactManager& checker)
   if (result_vector[0].link_names[0] != "sphere_link")
     idx = { 1, 0, -1 };
 
-  EXPECT_NEAR(result_vector[0].cc_time[idx[0]], 0.3333, 0.001);
-  EXPECT_NEAR(result_vector[0].cc_time[idx[1]], 0.5, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[0])], 0.3333, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[1])], 0.5, 0.001);
 
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[0])] == ContinuousCollisionType::CCType_Between);
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[1])] == ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[0]))] ==
+              ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[1]))] ==
+              ContinuousCollisionType::CCType_Between);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.05, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.05, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.05, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], -0.05, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][0], 0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][2], 0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][0], -0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][0], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][0], -0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_TRUE(result_vector[0].transform[idx[0]].isApprox(location_start["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].transform[idx[1]].isApprox(location_start["sphere1_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[0]].isApprox(location_end["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[1]].isApprox(location_end["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[0])].isApprox(location_start["sphere_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[1])].isApprox(location_start["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].cc_transform[static_cast<size_t>(idx[0])].isApprox(location_end["sphere_link"], 0.0001));
+  EXPECT_TRUE(
+      result_vector[0].cc_transform[static_cast<size_t>(idx[1])].isApprox(location_end["sphere1_link"], 0.0001));
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
   EXPECT_NEAR(result_vector[0].normal[1], idx[2] * 0.0, 0.001);
@@ -322,31 +328,34 @@ inline void runTestConvex(ContinuousContactManager& checker)
   if (result_vector[0].link_names[0] != "sphere_link")
     idx = { 1, 0, -1 };
 
-  EXPECT_NEAR(result_vector[0].cc_time[idx[0]], 0.5, 0.001);
-  EXPECT_NEAR(result_vector[0].cc_time[idx[1]], 0.5, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[0])], 0.5, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[1])], 0.5, 0.001);
 
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[0])] == ContinuousCollisionType::CCType_Between);
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[0])] == ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[0]))] ==
+              ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[0]))] ==
+              ContinuousCollisionType::CCType_Between);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.0377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.0377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.0377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], -0.0377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][0], 0.2377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][2], 0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][0], -0.2377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][0], 0.2377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][0], -0.2377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_TRUE(result_vector[0].transform[idx[0]].isApprox(location_start["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].transform[idx[1]].isApprox(location_start["sphere1_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[0]].isApprox(location_end["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[1]].isApprox(location_end["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[0])].isApprox(location_start["sphere_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[1])].isApprox(location_start["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].cc_transform[static_cast<size_t>(idx[0])].isApprox(location_end["sphere_link"], 0.0001));
+  EXPECT_TRUE(
+      result_vector[0].cc_transform[static_cast<size_t>(idx[1])].isApprox(location_end["sphere1_link"], 0.0001));
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
   EXPECT_NEAR(result_vector[0].normal[1], idx[2] * 0.0, 0.001);
@@ -390,31 +399,34 @@ inline void runTestConvex(ContinuousContactManager& checker)
   if (result_vector[0].link_names[0] != "sphere_link")
     idx = { 1, 0, -1 };
 
-  EXPECT_NEAR(result_vector[0].cc_time[idx[0]], 0.3848, 0.001);
-  EXPECT_NEAR(result_vector[0].cc_time[idx[1]], 0.5, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[0])], 0.3848, 0.001);
+  EXPECT_NEAR(result_vector[0].cc_time[static_cast<size_t>(idx[1])], 0.5, 0.001);
 
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[0])] == ContinuousCollisionType::CCType_Between);
-  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(idx[1])] == ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[0]))] ==
+              ContinuousCollisionType::CCType_Between);
+  EXPECT_TRUE(result_vector[0].cc_type[static_cast<size_t>(static_cast<size_t>(idx[1]))] ==
+              ContinuousCollisionType::CCType_Between);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][0], 0.0377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][1], 0.0772, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[0]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][0], 0.0377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][1], 0.0772, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][0], -0.0377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][1], 0.0772, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][0], -0.0377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][1], 0.0772, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][0], 0.2377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[0]][2], 0.25, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][0], -0.2377, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][1], 0.0, 0.001);
-  EXPECT_NEAR(result_vector[0].nearest_points_local[idx[1]][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][0], 0.2377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[0])][2], 0.25, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][0], -0.2377, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][1], 0.0, 0.001);
+  EXPECT_NEAR(result_vector[0].nearest_points_local[static_cast<size_t>(idx[1])][2], 0.25, 0.001);
 
-  EXPECT_TRUE(result_vector[0].transform[idx[0]].isApprox(location_start["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].transform[idx[1]].isApprox(location_start["sphere1_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[0]].isApprox(location_end["sphere_link"], 0.0001));
-  EXPECT_TRUE(result_vector[0].cc_transform[idx[1]].isApprox(location_end["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[0])].isApprox(location_start["sphere_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].transform[static_cast<size_t>(idx[1])].isApprox(location_start["sphere1_link"], 0.0001));
+  EXPECT_TRUE(result_vector[0].cc_transform[static_cast<size_t>(idx[0])].isApprox(location_end["sphere_link"], 0.0001));
+  EXPECT_TRUE(
+      result_vector[0].cc_transform[static_cast<size_t>(idx[1])].isApprox(location_end["sphere1_link"], 0.0001));
 
   EXPECT_NEAR(result_vector[0].normal[0], idx[2] * 1.0, 0.001);
   EXPECT_NEAR(result_vector[0].normal[1], idx[2] * 0.0, 0.001);
